Refactored controller concern to use loggable extension
